split the alignment functionality into several smaller focused modules
