+++ nc2pt/conf/climate_models/lr.yaml
@@ -19,7 +19,7 @@ hr_ref:
 
 climate_variables:
   # Activate variables to include in the LR model
-  #- ${internal.lr_uas}
+  - ${internal.lr_uas}
   #- ${internal.lr_vas}
   #- ${internal.lr_tas}
-  - ${internal.lr_pr}
+  #- ${internal.lr_pr}
