+++ nc2pt/conf/climate_models/lr_emulation/uas.yaml
@@ -5,8 +5,8 @@ name: "uas"
 alternative_names: ["U10", "u10", "uas"]
 path: ${internal.paths.lr.uas}
 is_west_negative: false
-apply_standardize: false
-apply_normalize: true
+apply_standardize: true
+apply_normalize: false
 invariant: false
 transform: []
 metadata_path: ${internal.paths.lr_emulation.uas_metadata}
