+++ nc2pt/conf/config.yaml
@@ -23,7 +23,7 @@ output_path: /home/sbeairsto/data-sbeairsto/config_test
 climate_models:
   #- ${internal.hr}
   #- ${internal.lr}
-  #- ${internal.lr_invariant}
-  #- ${internal.hr_invariant}
+  - ${internal.lr_invariant}
+  - ${internal.hr_invariant}
   - ${internal.lr_emulation}
   #- ${internal.my_model}
