+++ nc2pt/conf/paths.yaml
@@ -36,10 +36,10 @@ paths:
   # User-dependent path for – Low-Resolution ERA5 Emulation Metadata (LR Emulation) 
   # ────────────────────────────────────────
   lr_emulation:
-    uas_metadata: /home/sbeairsto/data-sbeairsto/config_test/uas_normalization_metadata.json
-    vas_metadata: /home/sbeairsto/data-sbeairsto/config_test/vas_normalization_metadata.json
-    pr_metadata:  /home/sbeairsto/data-sbeairsto/config_test/pr_normalization_metadata.json
-    tas_metadata: /home/sbeairsto/data-sbeairsto/config_test/tas_normalization_metadata.json
+    uas_metadata: /home/sbeairsto/data-sbeairsto/config_test/lr_uas_feature_scaling_metadata.json
+    vas_metadata: /home/sbeairsto/data-sbeairsto/config_test/lr_vas_feature_scaling_metadata.json
+    pr_metadata:  /home/sbeairsto/data-sbeairsto/config_test/lr_pr_feature_scaling_metadata.json
+    tas_metadata: /home/sbeairsto/data-sbeairsto/config_test/lr_tas_feature_scaling_metadata.json
 
   # ────────────────────────────────────────
   # User-dependent path for My Model Template
